Move `methods.rs` to it's own module one level above.
