+++ crates/ark/src/lib.rs
@@ -21,6 +21,7 @@ pub mod json;
 pub mod logger;
 pub mod logger_hprof;
 pub mod lsp;
+pub mod methods;
 pub mod modules;
 pub mod modules_utils;
 pub mod plots;
